Plots, Final Descents and Newest Writeup
Generated the final plots for the stability evaluation of the system. All that is to follow is a final run on the fine tuning of the descent. The writeup now contains all but one of the images that were planned. Still needs fixing of some of the concerns raised in the initial comments.